Skewed slicer
Made some minor changes to make code easier to use for other users. Also I am working on a "skewed" airfoil slicer; instead of having the root and chord planes parallel to the XY planes and potentially wasting a lot of foam with highly swept wings, this skews the wing planform to save foam. It does not work yet though.