Arduino code
The arduino is connected to a load cell via an HX711 amplifier and a stepper motor using an Allegro A4988 driver. This enables constant tension to be applied to the wire, regardless of how offset the toolheads are from each other. 

PID will have to be tuned depending on how rigidly the machine is constructed.